langgraph-agent/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.34,<0.35
langgraph>=0.3,<0.4
langchain-anthropic>=0.3,<0.4
psycopg[binary]>=3.2,<4
asyncpg>=0.30,<0.31
pydantic>=2.10,<3
pydantic-settings>=2.7,<3
email-validator>=2.0,<3
httpx>=0.28,<0.29
python-dotenv>=1.0,<2
# LangSmith tracing
langsmith>=0.2,<1
# Google APIs (calendar, sheets)
google-auth>=2.0,<3
